remove grove docker
